更新app/...

+++ pkg/timez/timez.go
@@ -3,6 +3,6 @@ package timez
 import "time"
 
 // UnixSecondNow 返回当前的 Unix 秒时间
-func UnixSecondNow() uint32 {
-	return uint32(time.Now().Unix())
+func UnixSecondNow() int64 {
+	return time.Now().Unix()
 }
